+++ CHANGELOG.md
@@ -9,6 +9,10 @@
 ### Fixed
 - block controller now gets correct data
 
+### Removed
+- Settings tab is not shown by default.
+    - Use `layouts/page-builder-block.yml` blueprint to override the default block blueprint.
+
 
 ## [1.0.2] - 2023-08-12
 ### Changed